Update preprocessing algorithm and move it to an explicit set of functions
Rather than leaving it implicitly in the data loading. The new way is to take
the mean value over each channel over all samples and shift the values to make
the mean 0, then scale each value by the standard deviation (taken in the same
way), so that e.g. the standard deviation of all of the prompt charges received
is 1, that of all of the prompt times is 1, etc.