Fix bug for retrieving cost in train_once (blame eb9c7c2a)
The cost used to be returned as the sole output of train_once, but then I added
the prediction also, turning the output into a list. When I removed the
prediction, I assumed the output would no longer be a list, but I was wrong. I
decided to keep the output as a list so that it would be easy to add new items
into the list later if desired, without requiring much other code modification.