Change implementation of corruption of DCA input to use dropout
This was on suggestion from Evan and is more elegant and simple than what I had
before (creating a mask and applying it to the input before supplying the input
to the network). My testing of this change revealed a bug in the network
implemenetation: it was not guaranteed to run on inputs smaller than the
specified minibatch size (supplied to the constructor as a user parameter). I
am willing to leave it that way, with the understanding that the user must
ensure the minibatch size is less than or equal to the total input size.